Cloud Computing/MainApp/requirements.txt
flask==3.0.*
functions-framework==3.*
google-cloud-bigquery==3.23.0
pandas==2.2.*
scikit-learn==1.5.*
gunicorn